Renames intercept.js to IncrementalDomAop.js to prevent blockers like uBlock origin from blocking it
