+++ packages/metal-incremental-dom/src/children/children.js
@@ -2,7 +2,7 @@
 
 import { buildCallFromConfig, buildConfigFromCall } from '../callArgs';
 import { isDef } from 'metal';
-import { startInterception, stopInterception } from '../intercept';
+import { startInterception, stopInterception } from '../IncrementalDomAop';
 
 /**
  * Property identifying a specific object as a Metal.js child node, and
