+++ packages/metal-incremental-dom/src/render/render.js
@@ -8,7 +8,7 @@ import { domData } from 'metal-dom';
 import { getData } from '../data';
 import { getCompatibilityModeData, getUid, isDef, isDefAndNotNull, isFunction, isString, object } from 'metal';
 import { disposeUnused, schedule } from '../cleanup/unused';
-import { getOriginalFn, startInterception, stopInterception } from '../intercept';
+import { getOriginalFn, startInterception, stopInterception } from '../IncrementalDomAop';
 import { Component, ComponentRegistry } from 'metal-component';
 
 const renderingComponents_ = [];
