+++ packages/metal-incremental-dom/test/IncrementalDomAop.js
@@ -2,9 +2,9 @@
 
 import core from 'metal';
 import dom from 'metal-dom';
-import { getOriginalFns, startInterception, stopInterception } from '../src/intercept';
+import { getOriginalFns, startInterception, stopInterception } from '../src/IncrementalDomAop';
 
-describe('intercept', function() {
+describe('IncrementalDomAop', function() {
 	var element;
 
 	beforeEach(function() {
